Add automatic CF invalidation

+++ dev/remove_old_files.sh
@@ -12,6 +12,10 @@ else
     exit 1
 fi
 
+if [[ -z $BUCKET ]]; then
+    exit
+fi
+
 for f in `aws s3 ls --recursive $BUCKET | awk '{print $4}'`; do
     if [[ ! -f dist/$f ]]; then
         aws s3 rm s3://$BUCKET/$f
